Major revamp to resolution process
Instead of trying to resolve all the problematic actions at once,
only apply the resolution to one action at a time.

To choose the action, an heuristic is used: For each problematic action,
look at the actions directly bellow, finding the corresponding actions
in the rest of the process. The higher that the corresponding actions
are in the process, the more likely unblocking the problematic action will
unblock the extra corresponding actions as well.